% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents for a given state in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number}

\item{year}{A number}
}
\value{
The plot of accidents for given \code{state.num} in \code{year}
}
\description{
Display accidents for a given state in a given year
}
\note{
if state.num not exist in data, the function will stop.
}
\examples{
\dontrun{fars_map_state(1, 2013)}

}
