% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a file at given path}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string}
}
\value{
The data frame from file at \code{filename}
}
\description{
Read data from a file at given path
}
\examples{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
fars_read("accident_2015.csv.bz2")
}
