% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get data for given year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An list of number}
}
\value{
The datafrome coresponding to given \code{years}
}
\description{
Get data for given year(s)
}
\examples{
fars_read_years(c(2013, 2015))
fars_read_years(c(2015,2018))
}
