% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data for given year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An list of number}
}
\value{
The data frame summarize data from given \code{years}
}
\description{
Summarize data for given year(s)
}
\examples{
fars_summarize_years(c(2013,2014))
fars_summarize_years(c(2015))
}
