% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct the filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number}
}
\value{
The string coresponding to \code{year}
}
\description{
Construct the filename for a given year
}
\examples{
\dontrun{make_filename(2018)}

}
