% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State Motor Vehicle Fatalities.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numerical code for US state.}

\item{year}{An integer, or a string or numeric that can be coerced to a string,
of the year of interest.}
}
\description{
\code{fars_map_state} maps state motor vehicle fatalities given a year and
  state id number.
}
\details{
This function takes a state number and a year, and draws
  a state outline with dots to represent the location of motor vehicle
  fatalities for that year.  This function will throw an error if an invalid
  state number is chosen or the chosen year's data does not exist.

You must have library(mapdata) loaded in your namespace for this to work.
}
\examples{
library(mapdata)
fars_map_state(12, 2014)
fars_map_state(36, 2014)

\dontrun{
fars_map_state(3, 2014)   # error
}

}

