% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename of the FARS data.}
}
\value{
This function returns a tibble containing the FARS data. If an
  incorrect filename is entered the function will stop.
}
\description{
\code{fars_read} reads FARS data into the environment.
}
\details{
This function reads data from the  US National Highway Traffic Safety
  Administration's Fatality Analysis Reporting System (FARS), given a
  filename for the data. It returns a tibble of the data. For this function
  to work properly, a filename pointing to an existing file must be given.
}
\examples{
full_filename <- system.file('extdata', 'accident_2013.csv.bz2',
                             package = 'farsdata')
fars_read(filename = full_filename)

\dontrun{
fars_read(filename = 'filedoesnotexist')
}

}

