% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a file at given path}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string}
}
\value{
The data frame from file at \code{filename}
}
\description{
Read data from a file at given path
}
\note{
The function will stop if the file does not exist.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
