% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS Filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, or a string or numeric that can be coerced to a string,
of the year of interest.}
}
\value{
this function returns a string that is the proper FARS data
   filename for the given year.  Will return a filename with NA for the
   year slot in the name if the year parameter cannot be coerced to an
   integer.
}
\description{
\code{make_filename} makes a properly formatted FARS filename given a year
  as input.
}
\details{
This function takes a year as input and produces a valid FARS filename. This
  function is "dumb" in that it will not check whether the file
  actually exists, or if the year is reasonable -- you could enter any number
  you like.
}
\examples{
make_filename(year = '2013')
make_filename(year = 2013)

\dontrun{
make_filename(year = 'two thousand thirteen') #  error
}


}

